Updated skill list and about me based on a recent change in projects.
Removed Go reference, no longer actively working with this.
Added Rust mention in favourites as it is my 'happy programming language'.

Added an explanation what this document is about.

Removed AI based story references - based on data it is trained on, I don't find it ethical. Only mentions are communication based which don't have a focus on generating stuff for you, rather teaching you how things work in a conversation.